refactor: extract tryInstallAppOnDevice
